Задание SDJpa.

+++ 2022-05/spring-07-jdbc/jdbc-demo-exercise/src/main/java/ru/otus/spring/service/ServiceBooks.java
@@ -16,7 +16,7 @@ public interface ServiceBooks {
 
     List<Books> getAll();
 
-    void deleteById(int number);
+    void deleteById(long number);
     public void bPrint(List<Books> books);
 
 }
