+++ 2022-05/spring-07-jdbc/jdbc-demo-exercise/src/main/java/ru/otus/spring/service/ServiceComments.java
@@ -9,7 +9,7 @@ public interface ServiceComments {
 
     Long count();
 
-    Optional<Comment> getById(int id);
+    Optional<Comment> getById(long id);
 
     List<Comment> getAll();
